Extract btstorage methods
