+++ freqtrade/optimize/optimize_reports/__init__.py
@@ -7,10 +7,11 @@ from freqtrade.optimize.optimize_reports.bt_output import (show_backtest_result,
                                                            text_table_exit_reason,
                                                            text_table_periodic_breakdown,
                                                            text_table_strategy, text_table_tags)
+from freqtrade.optimize.optimize_reports.bt_storage import (store_backtest_analysis_results,
+                                                            store_backtest_stats)
 from freqtrade.optimize.optimize_reports.optimize_reports import (
     generate_all_periodic_breakdown_stats, generate_backtest_stats, generate_daily_stats,
     generate_edge_table, generate_exit_reason_stats, generate_pair_metrics,
     generate_periodic_breakdown_stats, generate_rejected_signals, generate_strategy_comparison,
     generate_strategy_stats, generate_tag_metrics, generate_trade_signal_candles,
-    generate_trading_stats, generate_wins_draws_losses, store_backtest_analysis_results,
-    store_backtest_stats)
+    generate_trading_stats, generate_wins_draws_losses)
